feat: refactor and prep for animation

+++ config/boards/shields/custom_nice_view/include/utils/draw_background.h
@@ -2,4 +2,4 @@
 
 #include <lvgl.h>
 
-void draw_background(lv_obj_t* canvas, uint8_t index);
+void draw_background(lv_obj_t* canvas, unsigned index);
